requirements.txt
streamlit
pandas
numpy
msoffcrypto-tool
openpyxl
xlrd
